Restrict reversible check to a glob
Currently reversible check is applied to all the files of the
migration. Sometimes we have generated files or files that we
know are not reversible (For example when scrubbing files).

We still would like to check that we can reverse the transformations
for the rest of files.

This change adds an optional glob field that allow ignore some files
in the check.

An alternative considered was to use a reversible_check_files field,
that would be the negation of this one. The main problem with that
approach is that we expect that the files to be ignored to be
the exception. So it is easier to build a glob for files to ignore
than files to match (Specially as we don't have '-' opearator for
globs, so we cannot do origin_files - glob(["some_files"]). The
only issue with this approach is that the exclude field in the glob
would be the double negation, not ideal, but hopefully it is not
used.

BUG=149343687

Change-Id: I819040d5cbdad7f5a9502872e5c5727494effbff