Allow custom RevId
Currently revId is hardcoded per origin. There are some limitations on
how we can mix different imports to the same destination repo.
Currently it only works for repos like this:

project1/**
project2/**

We cannot be more strict with the glob for destinationStatus because it can
mutate and it wouldn't match old changes.

This change allows to customize the revid to separate changes. For now it is
an experiment for updating external dependencies.

PiperOrigin-RevId: 228915118
Change-Id: Ic7ce3a8bddf57b4a6cc3a4a07f06b23fb2170045