Externalize buildifier transformation
Buildifier has been open sourced, this open sources code to call it
from copybara. Using the new transform requires installing buildifier.

Shoutout to Yannic, who prompted us to release this in PR google#111. We have users relying on the signature to be as-is, but future work like adding the "mode" from the PR to turn this into a check sounds promising.

Closes google#111

PiperOrigin-RevId: 294962888
Change-Id: Ie38615d9383c8b03540953498481c37b54fd3944